add new class Calendar
