+++ src/module/utils.js
@@ -1,13 +1,28 @@
-export function row(content, style = '') {
-  return `<div class="row" style="${style}">${content}</div>`;
-}
+import { workWeek } from './auxiliary';
 
-export function col(content) {
-  return `<div class="col-sm">${content}</div>`;
-}
+export const capitalize = string => {
+  if (typeof string !== 'string') {
+    return '';
+  }
+  return string.charAt(0).toUpperCase() + string.slice(1);
+};
 
-export function css(styles = {}) {
-  if (typeof(styles) === 'string') { return styles; }
-  const toString = key => `${key}: ${styles[key]}`;
-  return Object.keys(styles).map(toString).join(';');
-}
+export const firstFreeSlot = calendar => {
+  let nextFreeTime = '';
+  let lastBookedDay = '';
+  workWeek.forEach(day => {
+    calendar.querySelectorAll(`.cell[data-day="${day.id}"]`)
+      .forEach(elem => {
+        if (!elem.dataset.eventId || elem.dataset.eventId.trim() === '') {
+          if (lastBookedDay === '' && nextFreeTime === '') {
+            lastBookedDay = elem.dataset.day;
+            nextFreeTime = elem.dataset.time;
+          }
+        }
+      });
+  });
+  return {
+    day: lastBookedDay,
+    time: nextFreeTime,
+  };
+}; // end firstFreeSlot
